% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_proximal_metabolites.R
\name{extract_proximal_metabolites}
\alias{extract_proximal_metabolites}
\title{Extract proximal metabolites from RaMP knowledge graph}
\usage{
extract_proximal_metabolites(
  metabolites,
  crosstalkers_full_set,
  cutoff = NULL,
  percentile_cutoff = NULL,
  absolute_cutoff = NULL,
  biospecimen = NULL,
  pathwaysOrChemicals = "pathways",
  keep_seeds = FALSE,
  conpass = NULL,
  dbname = "ramp",
  username = "root",
  host = "localhost",
  socket = NULL
)
}
\arguments{
\item{metabolites}{A list of database identifiers associated with metabolites of interest. Database of origin must be prepended to identifier (e.g. hmdb:HMDB0000001)}

\item{crosstalkers_full_set}{The output from "measure_metabolite_distance": a vector of distances between each node in RaMP and the seed set}

\item{cutoff}{Distance cutoff for extracted metabolites. "visualize_metabolite_distances" is helpful for determining a cutoff}

\item{percentile_cutoff}{Extracts top n percentile metabolites by lowest distance}

\item{absolute_cutoff}{Extracts top n metabolites by lowest distance}

\item{biospecimen}{If NULL, test all metabolites in RaMP. Else, test networks built for specific biospecimens. Choices are "Blood", "Adipose", "Heart", "Urine", "Brain", "Liver", "Kidney", "Saliva", and "Feces"}

\item{pathwaysOrChemicals}{Test distance on a pathway similarity or chemical similarity matrix. Choices are "pathways" or "chemicals"}

\item{keep_seeds}{Boolean for whether or not seed distances from input set should be calculated (almost always FALSE)}

\item{conpass}{Password for database access (string)}

\item{dbname}{Name of the mysql database (default is "ramp")}

\item{username}{Username for database access (default is "root")}

\item{host}{Host name for database access (default is "localhost")}

\item{socket}{(Optional) location of mysql.sock file}
}
\value{
Vector of metabolites proximal to the seed set that met specified cutoff criterion
}
\description{
Uses output from "measure_metabolite_distance" to assemble vector of proximal metabolites to the set of interest
}
\author{
Andrew Christopher Patt
}
