% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_metabolite_distance.R
\name{measure_metabolite_distance}
\alias{measure_metabolite_distance}
\title{Compute distance to seed set for each node in RaMP knowledge graph}
\usage{
measure_metabolite_distance(
  metabolites,
  biospecimen = NULL,
  pathwaysOrChemicals = "pathways",
  keep_seeds = FALSE,
  conpass = NULL,
  dbname = "ramp",
  username = "root",
  host = "localhost",
  socket = NULL
)
}
\arguments{
\item{metabolites}{A list of database identifiers associated with metabolites of interest. Database of origin must be prepended to identifier (e.g. hmdb:HMDB0000001)}

\item{biospecimen}{If NULL, test all metabolites in RaMP. Else, test networks built for specific biospecimens. Choices are "Blood", "Adipose", "Heart", "Urine", "Brain", "Liver", "Kidney", "Saliva", and "Feces"}

\item{pathwaysOrChemicals}{Test distance on a pathway similarity or chemical similarity matrix. Choices are "pathways" or "chemicals"}

\item{keep_seeds}{Boolean for whether or not seed distances from input set should be calculated (almost always FALSE)}

\item{conpass}{Password for database access (string)}

\item{dbname}{Name of the mysql database (default is "ramp")}

\item{username}{Username for database access (default is "root")}

\item{host}{Host name for database access (default is "localhost")}

\item{socket}{(Optional) location of mysql.sock file}
}
\value{
Vector of distances from each node in RaMP knowledge network to seed set
}
\description{
Uses Markov random walk with restarts algorithm to generate a by-node distance measurement from each node in the RaMP knowledge graph to each node associated with the input metabolites.
}
\author{
Andrew Christopher Patt
}
