% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_metabolite_network.R
\name{source_to_common}
\alias{source_to_common}
\title{Find rampId from given common name vector}
\usage{
source_to_common(
  source_ids,
  conpass = NULL,
  dbname = "ramp",
  username = "root",
  host = "localhost",
  socket = NULL
)
}
\arguments{
\item{source_ids}{a vector of metabolite common name strings}

\item{conpass}{password for database access (string)}

\item{dbname}{name of the mysql database (default is "ramp")}

\item{username}{username for database access (default is "root")}

\item{host}{host name for database access (default is "localhost")}

\item{socket}{(optional) location of mysql.sock file}
}
\value{
data.frame that has rampId as column
}
\description{
Find rampId from given common name vector
}
