% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_metabolite_distances.R
\name{visualize_metabolite_distances}
\alias{visualize_metabolite_distances}
\title{Visualize the distance distribution between nodes in the RaMP network and metabolites of interest}
\usage{
visualize_metabolite_distances(crosstalkers_full_set)
}
\arguments{
\item{crosstalkers_full_set}{The output from "measure_metabolite_distance": a vector of distances between each node in RaMP and the seed set}
}
\value{
Histogram of distance distribution (plot)
}
\description{
Generate a histogram of distance measurements from each node in the RaMP knowledge network to the metabolite set of interest (uses the output from measure_metabolite_distance)
}
\author{
Andrew Christopher Patt
}
