% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_metabolite_network.R
\name{visualize_metabolite_network}
\alias{visualize_metabolite_network}
\title{Internal function for visualizing metabolite network}
\usage{
visualize_metabolite_network(
  metabolites,
  crosstalkers,
  biospecimen = NULL,
  pathwaysOrChemicals = "pathways",
  label = TRUE,
  filterIslands = TRUE,
  conpass = NULL,
  dbname = "ramp",
  username = "root",
  host = "localhost",
  socket = NULL
)
}
\arguments{
\item{metabolites}{A list of database identifiers associated with metabolites of interest. Database of origin must be prepended to identifier (e.g. hmdb:HMDB0000001)}

\item{crosstalkers}{Vector of metabolites returned by "extract_proximal_metabolites"}

\item{biospecimen}{If NULL, test all metabolites in RaMP. Else, test networks built for specific biospecimens. Choices are "Blood", "Adipose", "Heart", "Urine", "Brain", "Liver", "Kidney", "Saliva", and "Feces"}

\item{pathwaysOrChemicals}{Test distance on a pathway similarity or chemical similarity matrix. Choices are "pathways" or "chemicals"}

\item{label}{Boolean specifying if nodes in visual should be labeled with metabolite common names}

\item{filterIslands}{Boolean to display "Islands" (nodes with no direct connections to other nodes in visual)}

\item{conpass}{Password for database access (string)}

\item{dbname}{Name of the mysql database (default is "ramp")}

\item{username}{Username for database access (default is "root")}

\item{host}{Host name for database access (default is "localhost")}

\item{socket}{(Optional) location of mysql.sock file}
}
\value{
Visual showing connectivity of seeds and extracted nodes from data
}
\description{
Shows seed metabolites, extracted metabolites, and their connectivity in the RaMP knowledge graph
}
\author{
Andrew Christopher Patt
}
